PIM-6175: Move UserBundle repository to right folders
