+++ src/Pim/Bundle/CatalogBundle/Resources/config/savers.yml
@@ -97,7 +97,5 @@ services:
             - '@event_dispatcher'
             - '%pim_catalog.model.locale.interface%'
 
-    # deprecated, will be removed in 1.6 (use akeneo_storage_utils.saver.base_options_resolver instead)
-
     pim_catalog.saver.group_options_resolver:
         class: '%pim_catalog.saver.group_options_resolver.class%'
