+++ src/Pim/Bundle/UserBundle/Provider/UserProvider.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\UserBundle\Provider;
 
 use Doctrine\Common\Util\ClassUtils;
-use Pim\Bundle\UserBundle\Entity\Repository\UserRepositoryInterface;
+use Pim\Bundle\UserBundle\Repository\UserRepositoryInterface;
 use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
 use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
 use Symfony\Component\Security\Core\User\UserInterface;
